requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
argh==0.26.2              # via watchdog
astroid==1.6.5            # via pylint
isort==4.3.4              # via pylint
lazy-object-proxy==1.3.1  # via astroid
mccabe==0.6.1             # via pylint
pathtools==0.1.2          # via watchdog
pillow==5.2.0
pylint==1.9.2
pyyaml==3.13              # via watchdog
six==1.11.0               # via astroid, pylint
watchdog==0.8.3
wrapt==1.10.11            # via astroid
